refactor: 이미지 라이선스 컨트롤러 통합
